requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
numpy==1.26.4
pillow==11.0.0
tensorflow==2.17.0
matplotlib==3.9.2
python-multipart==0.0.12
